Diagnose and fix a real regression: 2D model scored worse than bicubic
A full inspection + quantitative baseline (scripts/evaluate.py, new) found
that the previous 2D super-resolution checkpoint measurably underperformed
plain bicubic upscaling on held-out real imagery (PSNR/SSIM both worse,
confirmed on a genuinely unseen test image never used in training).
Visual inspection showed the model hallucinating noise-like texture into
smooth regions (open water) its 2-image training corpus barely contained.

Root causes fixed:
- train_2d.py only ever saved the final epoch, with no validation signal
  to tell whether it generalized. It now performs a spatial train/val
  split per source image (SatelliteSRDataset gained split/val_fraction)
  and tracks per-epoch held-out PSNR/SSIM, saving the best-scoring
  checkpoint separately instead of trusting the last epoch.
- The composite loss's hand-rolled "perceptual" edge term measurably
  encouraged texture hallucination at this dataset size; its weight now
  defaults to 0 (configurable via --weight-perceptual).
- metrics.psnr() returned a numpy scalar instead of a native float, which
  broke torch.load(weights_only=True) once that value got stored in a
  checkpoint dict -- a separate real bug found while wiring up validation.
- infer_2d.py's tiled inference used flat uniform-weight blending across
  overlapping tiles, a known source of visible seams; replaced with a
  feathered (ramped) blend weighted toward each tile's higher-context
  interior.

Also:
- Added a third real training image (a genuine USGS aerial/terrain scene)
  and a fourth real image held out strictly for evaluation, documenting
  in scripts/download_sample_data.py which other "sample data" sources
  turned out to be synthetic noise or mislabeled single-band rasters.
- tests/test_regression.py guards against this exact failure mode going
  forward (verified it fails against the old checkpoint before trusting
  it as a real regression test).
- README/ARCHITECTURE.md document the diagnosis, the fix, and known
  limitations honestly instead of asserting the system just works.

Training with the new methodology is running; the final checkpoint will
follow in a separate commit once validated against scripts/evaluate.py.

+++ src/satellite_enhance/metrics.py
@@ -10,7 +10,7 @@ def psnr(pred: np.ndarray, target: np.ndarray, data_range: float = 1.0) -> float
     mse = float(np.mean((pred.astype(np.float64) - target.astype(np.float64)) ** 2))
     if mse == 0:
         return float("inf")
-    return 10.0 * np.log10((data_range**2) / mse)
+    return float(10.0 * np.log10((data_range**2) / mse))
 
 
 def ssim(pred: np.ndarray, target: np.ndarray, data_range: float = 1.0) -> float:
